LayoutTests/fast/dom/MicroData/itemref-attribute-reflected-by-itemRef-property.html

<!DOCTYPE html>
<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
<script src="resources/microdata-common.js"></script>
</head>
<body>
<p>This test ensures that the itemref attribute must be reflected by the .itemRef property.</p>
<div id="console"></div>
<script>
var element = createElement('div', {itemref: 'foo bar'});
shouldBeDefined("element.itemRef");

shouldBeTrue("element.itemRef.length == 2");
shouldBeTrue("element.itemRef.toString() == 'foo bar'");

element.removeAttribute('itemref');
shouldBeTrue("element.itemRef.toString() == ''");

</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
